requirements.txt
pandas==1.5.2
psycopg2==2.9.5
pymysql==1.0.2
cryptography==37.0.4
PyYAML==6.0
sqlalchemy